Add clang-format rules and makefile targets
Will commit the result in the next changeset.

Also, had to update to a newer version of LLVM to get some new clang-format options.
We don't necessarily have to use the same version of clang-format as the version of
llvm that we build it against, but if we can it keeps the build system simpler.