More makefile magic
This time for the llvm configuration.

Sometimes when updating the LLVM directory, we need to reconfigure things.
We don't know that until after we update it, though, but Make will decide
at the beginning whether it needs to be updated or not.  We could always
reconfigure every time we update LLVM (might be ok), or add a helper
target that recursively calls Make after the updating is done (what this
patch does).

This should fix those "C compiler can't generate executables" errors that
come from running two copies of the configure script in parallel.