Raise this as an ImportError instead of SystemError
With the recent package import support, we broke importing optparse.
This is because we now can import "encodings" (a package), which imports
codecs, which raises a SystemError when trying to import _codecs which
we don't support yet.

This was getting ignored before since optparse wrapped this process
(which happens via "import gettext") in a try-except and catches the
ImportError.