Modify the stdlib to get distutils working
There are two tricky things left that I don't think we can support soon:
- MultipartConversionError uses multiple inheritance
- structseq.c sets a tp_dealloc

The structseq_dealloc function is a "simple" destructer (functionality is
covered by our GC), so we can just comment it out.  Ideally we'd have some
way of automatically determining that it's a no-op so that we don't have
to run it.