Rebase to LLVM r223801
Primary challenge is rebasing past the JITEventListener changes.

Some API changes, but also some considerable functionality changes,
since we no longer get the loaded version of the object file.

They added a feature to get the load address of a section, but not of
a symbol; I think that makes sense so I'll submit a patch for that.