Push back the performance regression
The previous changes increased -O execution time, showing up mostly in
definedness analysis and type analysis

I guess that makes sense, since we're adding additional symbols that
we then have to analyze.  We should make the analyses faster in general.

For now, just eliminate some of the unnecessary symbols that we're generating.
We only need to create the NONZERO temporary if we need to care about exception
state, so if there's no exception handler, just skip the temporary and generate
a nested structure.