Implement Python's name mangling
ie names that begin with two underscores but don't end in two underscores have
the classname added to them.

Do this early on in the pipeline so that all the analyses operate post-mangling.

The implementation is kind of hacky and I couldn't think of a good way to make it
super systematic; there may be more cases I missed.