Scan ht_name for heap types
We weren't scanning it as part of the gc handler, so eventually it would
get collected and then __name__ would break.  (un)fortunately, in the
cases that would use it, it looks like it would always get replaced by
another string (the only thing being allocated out of that size bin?),
so it would look like the __name__ was just wrong.

Anyway, this should fix interp2.py