Add CAPI-style getset descriptors as well
We already have getset support, but it takes function pointers
to Pyston-style functions.  CAPI types can specify tp_getset which
is almost the same thing, but the function pointers are to CAPI-style
functions.

Add the ability to do either, try to make it more explicit about
which one is being chosen, and change the existing getters/setters to be
more consistent with being Pyston-style functions.