Fix the caching for the file test
The file.py test looked at the README.md file, but the tester caches
the CPython output and when I changed the README it didn't invalidate the
cache.

The tester knows to invalidate the cache if the test file itself changes, so
instead of using README as the test target, use itself since that dependency
is already tracked.