Fix some importing mechanics that particularly affect import hooks
There are some times that the importing logic has to look up attributes
on objects, and we were using Box::getattr() (equivalent to
obj.__dict__[attr] ) instead of getattr() (equivalent to getattr()).
This was working fine until we started trying to run the custom import hook
in six.py, since that creates a subclass of ModuleType which defines
some of the important attributes as class-level attributes.