Allow some more use of unicode as attribute or variable names
Treat unicode names as equivalent to their ascii-encoded str
version, and any non-ascii unicode name as a TypeError.

My hope is that this covers the common cases but doesn't cause
bugs.  You can see that we're doing this by putting unicode in
and getting a str out, but I hope that this is ok.