add a json test and get it running
there are a few areas where we fail (decimal usage, and formatting complex numbers)
so those specific lines are commented out.

The changes in src were necessary to get the test to run to completion with -n.
Finally, we skip the test when -x is present because pypa doesn't parse unicode
literals correctly.