Don't try to specialize when interpreting.
1) because we haven't seen the types multiple times yet,
so creating a new function specialization could be wasteful, and
2) the interpreter can't make use of the type specialization.

Specializing at the interpreter does let us start counting how
many times each specialization is used, and the function entry
counts are the only way we have right now of telling that a
particular specialization should be moved to a higher tier.
But I don't think we should do that using this mechanism.